+++ TwitchLib.Client.Tests/Services/ThrottlerServiceTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+using Xunit;
+
+namespace TwitchLib.Client.Tests.Services
+{
+    public class ThrottlerServiceTests
+    {
+        [Fact]
+        public void DoSomeTests() { throw new NotImplementedException(); }
+    }
+}