+++ TwitchLib.Client.Tests/TwitchClient_ChannelTests.cs
@@ -60,7 +60,7 @@ namespace TwitchLib.Client.Tests
                         Task.Delay(2000).GetAwaiter().GetResult();
                         // we have to cheat a bit
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheckInitial.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertionInitial.Arguments);
@@ -82,7 +82,7 @@ namespace TwitchLib.Client.Tests
                         client.OnChannelStateChanged += (sender, args) => Assert.True(pauseCheckUpdate.Set());
                         // we have to cheat a bit
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheckUpdate.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertionUpdate.Arguments);
@@ -104,14 +104,12 @@ namespace TwitchLib.Client.Tests
             // to make the ITwitchClient call ChannelManager.Start()
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":tmi.twitch.tv 004 {TWITCH_Username} :-" });
 
 
             string message = $"@msg-id={MsgIds.MsgChannelSuspended} :tmi.twitch.tv NOTICE #{TWITCH_CHANNEL} :This channel does not exist or has been suspended.";
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = message });
 
             IClient communicationClient = mock.Object;
@@ -130,7 +128,7 @@ namespace TwitchLib.Client.Tests
                         // make the client raise OnConnected and ITwitchClient start ChannelManager
                         communicationClient.Send(String.Empty);
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -147,14 +145,12 @@ namespace TwitchLib.Client.Tests
             // to make the ITwitchClient call ChannelManager.Start()
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":tmi.twitch.tv 004 {TWITCH_Username} :-" });
 
             // after calling IClient.JoinChannel([...]), the ChannelManager should send "JOIN #[...]"
             // and we want to raise the Join-Confirmation-Message, that we recveive from twitch
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.Is<string>(s => String.Equals($"JOIN #{TWITCH_CHANNEL}", s))))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":{TWITCH_Username}!{TWITCH_Username}@{TWITCH_Username}.tmi.twitch.tv JOIN #{TWITCH_CHANNEL}" });
 
             IClient communicationClient = mock.Object;
@@ -190,20 +186,17 @@ namespace TwitchLib.Client.Tests
             // to make the ITwitchClient call ChannelManager.Start()
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":tmi.twitch.tv 004 {TWITCH_Username} :-" });
 
             // after calling IClient.JoinChannel([...]), the ChannelManager should send "JOIN #[...]"
             // and we want to raise the Join-Confirmation-Message, that we recveive from twitch
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.Is<string>(s => String.Equals($"JOIN #{TWITCH_CHANNEL}", s))))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":{TWITCH_Username}!{TWITCH_Username}@{TWITCH_Username}.tmi.twitch.tv JOIN #{TWITCH_CHANNEL}" });
 
             // 
             mock.InSequence(sequence)
                 .Setup(c => c.Send(It.Is<string>(s => String.Equals($"PART #{TWITCH_CHANNEL}", s))))
-                .Returns(true)
                 .Raises(c => c.OnMessage += null, new OnMessageEventArgs() { Message = $":{TWITCH_Username}!{TWITCH_Username}@{TWITCH_Username}.tmi.twitch.tv PART #{TWITCH_CHANNEL}" });
 
             IClient communicationClient = mock.Object;