+++ TwitchLib.Client.Tests/TwitchClient_ConnectionTests.cs
@@ -90,7 +90,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnConnected += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -124,7 +124,7 @@ namespace TwitchLib.Client.Tests
                         () =>
                         {
                             // send is our trigger, to make the IClient-Mock raise OnMessage!
-                            Assert.True(communicationClient.Send(String.Empty));
+                            communicationClient.Send(String.Empty);
                             // we dont need to wait to long, we expect it to fail
                             Assert.False(pauseCheck.WaitOne(WaitOneDurationShort));
                         });
@@ -179,7 +179,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnIncorrectLogin += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -191,7 +191,6 @@ namespace TwitchLib.Client.Tests
         {
             Mock<IClient> mock = IClientMocker.GetIClientMock();
             mock.Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(false)
                 .Raises(c => c.OnFatality += null, new OnFatalErrorEventArgs("Fatal network error."));
             IClient communicationClient = mock.Object;
             // create one logger per test-method! - cause one file per test-method is generated