+++ TwitchLib.Client.Tests/TwitchClient_MessageSendingTests.cs
@@ -11,7 +11,6 @@ using TwitchLib.Client.Events;
 using TwitchLib.Client.Interfaces;
 using TwitchLib.Client.Models;
 using TwitchLib.Client.Tests.TestHelpers;
-using TwitchLib.Communication.Events;
 using TwitchLib.Communication.Interfaces;
 
 using Xunit;
@@ -67,7 +66,7 @@ namespace TwitchLib.Client.Tests
                         Task.Delay(2000).GetAwaiter().GetResult();
                         // we have to cheat a bit
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheckJOIN.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertionInitial.Arguments);
@@ -85,7 +84,7 @@ namespace TwitchLib.Client.Tests
                         client.OnMessageSent += (sender, args) => Assert.True(pauseCheckUpdate.Set());
                         // we have to cheat a bit
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheckUpdate.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertionUpdate.Arguments);
@@ -97,49 +96,42 @@ namespace TwitchLib.Client.Tests
             uint sendsAllowedInPeriod = 0;
             TimeSpan throttlingTimeSpan = TimeSpan.FromSeconds(30);
             string messageNotSent = "This message has not been sent!";
-            string reason = "Some reason";
-
-            OnThrottledEventArgs onThrottledEventArgs = new OnThrottledEventArgs()
-            {
-                AllowedInPeriod = sendsAllowedInPeriod,
-                SentCount = sendsAllowedInPeriod,
-                Period = throttlingTimeSpan,
-                ItemNotSent = messageNotSent,
-                Reason = reason
-            };
-
-            OnMessageThrottledEventArgs onMessageThrottledEventArgs = new OnMessageThrottledEventArgs(onThrottledEventArgs);
-
-            Mock<IClient> clientMock = new Mock<IClient>();
-
-            clientMock.SetupAdd(c => c.OnMessageThrottled += It.IsAny<EventHandler<OnMessageThrottledEventArgs>>());
-            clientMock.Setup(c => c.Send(It.IsAny<string>()))
-                .Returns(false)
-                .Raises(c => c.OnMessageThrottled += null, onMessageThrottledEventArgs);
-
-            IClient communicationClient = clientMock.Object;
+            Mock<IClient> mock = IClientMocker.GetIClientMock();
+            MockSequence sendMessageSequnce = new MockSequence();
+            IClientMocker.AddLogInToSendMessageSequence($":tmi.twitch.tv 004 {TWITCH_Username} :-", mock, sendMessageSequnce);
+            IClientMocker.AddToSendMessageSequence($":{TWITCH_Username}!{TWITCH_Username}@{TWITCH_Username}.tmi.twitch.tv JOIN #{TWITCH_CHANNEL}", mock, sendMessageSequnce);
+            IClient communicationClient = mock.Object;
 
             // create one logger per test-method! - cause one file per test-method is generated
             ILogger<ITwitchClient> logger = TestLogHelper.GetLogger<ITwitchClient>();
             ConnectionCredentials connectionCredentials = new ConnectionCredentials(TWITCH_Username, TWITCH_OAuth);
-            ITwitchClient client = new TwitchClient(connectionCredentials, communicationClient, logger: logger);
+            ITwitchClient client = new TwitchClient(credentials: connectionCredentials,
+                                                    client: communicationClient,
+                                                    sendOptions: new SendOptions(sendsAllowedInPeriod),
+                                                    logger: logger);
+            client.JoinChannel(TWITCH_CHANNEL);
             ManualResetEvent pauseCheck = new ManualResetEvent(false);
-            Assert.RaisedEvent<OnMessageThrottledEventArgs> assertion = Assert.Raises<OnMessageThrottledEventArgs>(
+            Assert.RaisedEvent<OnMessageThrottledArgs> assertion = Assert.Raises<OnMessageThrottledArgs>(
                     h => client.OnMessageThrottled += h,
                     h => client.OnMessageThrottled -= h,
                     () =>
                     {
                         client.OnMessageThrottled += (sender, args) => Assert.True(pauseCheck.Set());
-                        communicationClient.Send(messageNotSent);
+                        IClientMocker.SetIsConnected(mock, true);
+                        client.Connect();
+                        // wait to connect and join
+                        Task.Delay(2000).GetAwaiter().GetResult();
+                        client.SendMessage(TWITCH_CHANNEL, messageNotSent);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
             Assert.NotNull(assertion.Arguments.Reason);
-
-            Assert.Equal(messageNotSent, assertion.Arguments.ItemNotSent);
-            Assert.Equal(reason, assertion.Arguments.Reason);
+            Assert.NotNull(assertion.Arguments.ItemNotSent);
+            Assert.Equal(messageNotSent, assertion.Arguments.ItemNotSent.Message);
             Assert.Equal(throttlingTimeSpan, assertion.Arguments.Period);
             Assert.Equal(sendsAllowedInPeriod, assertion.Arguments.AllowedInPeriod);
         }
+        [Fact]
+        public void TwitchClient_Raises_OnSendFailed() { throw new NotImplementedException(); }
     }
 }