+++ TwitchLib.Client.Tests/TwitchClient_NoticeUserTests.cs
@@ -34,7 +34,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnAnnouncement += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -60,7 +60,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnRaidNotification += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -86,7 +86,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnUnRaidNotification += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -112,7 +112,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnNewSubscriber += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -138,7 +138,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnReSubscriber += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -164,7 +164,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnPrimePaidSubscriber += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -190,7 +190,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnGiftedSubscription += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -216,7 +216,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnCommunitySubscription += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);
@@ -242,7 +242,7 @@ namespace TwitchLib.Client.Tests
                     {
                         client.OnContinuedGiftedSubscription += (sender, args) => Assert.True(pauseCheck.Set());
                         // send is our trigger, to make the IClient-Mock raise OnMessage!
-                        Assert.True(communicationClient.Send(String.Empty));
+                        communicationClient.Send(String.Empty);
                         Assert.True(pauseCheck.WaitOne(WaitOneDuration));
                     });
             Assert.NotNull(assertion.Arguments);