+++ TwitchLib.Client/TwitchClient.cs
@@ -11,6 +11,7 @@ using TwitchLib.Client.Extensions.Internal;
 using TwitchLib.Client.Interfaces;
 using TwitchLib.Client.Managers;
 using TwitchLib.Client.Models;
+using TwitchLib.Client.Models.Interfaces;
 using TwitchLib.Communication.Clients;
 using TwitchLib.Communication.Interfaces;
 
@@ -39,9 +40,48 @@ namespace TwitchLib.Client
 
 
         #region ctor
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="credentials">
+        ///     <inheritdoc cref="ConnectionCredentials"/>
+        ///     <br></br>
+        ///     see
+        ///     <br></br>
+        ///     <see cref="Models.ConnectionCredentials"/>
+        /// </param>
+        /// <param name="client">
+        ///     <inheritdoc cref="IClient"/>
+        ///     <br></br>
+        ///     see
+        ///     <br></br>
+        ///     <see cref="IClient"/>
+        /// </param>
+        /// <param name="protocol">
+        ///     <inheritdoc cref="ClientProtocol"/>
+        ///     <br></br>
+        ///     see
+        ///     <br></br>
+        ///     <see cref="ClientProtocol"/>
+        /// </param>
+        /// <param name="sendOptions">
+        ///     <inheritdoc cref="ISendOptions"/>
+        ///     <br></br>
+        ///     by leaving it <see langword="null"/>,
+        ///     <see langword="default"/> <see cref="SendOptions"/>
+        ///     with the minimum <see cref="MessageRateLimit.Limit_20_in_30_Seconds"/>
+        ///     is going to be applied
+        /// </param>
+        /// <param name="logger">
+        ///     an <see cref="ILogger"/>
+        /// </param>
+        /// <exception cref="ArgumentNullException">
+        ///     if <paramref name="credentials"/> is <see langword="null"/>
+        /// </exception>
         public TwitchClient(ConnectionCredentials credentials,
                             IClient client = null,
                             ClientProtocol protocol = ClientProtocol.WebSocket,
+                            ISendOptions sendOptions = null,
                             ILogger<ITwitchClient> logger = null)
         {
             LOGGER = logger;
@@ -75,6 +115,12 @@ namespace TwitchLib.Client
             // cause credentials are also set into ChannelManager
             SetConnectionCredentials(credentials);
             ChatCommandIdentifiers.Add('!');
+            if (sendOptions == null) sendOptions = new SendOptions((uint) MessageRateLimit.Limit_20_in_30_Seconds);
+            ThrottlerService = new Services.ThrottlerService(Client,
+                                                             this,
+                                                             sendOptions,
+                                                             LOGGER);
+            ThrottlerService.Subscribe(this);
         }
         #endregion ctor